Much changes -- allow PDF to be very flexibly described in INFILE. More interactive mode stuff. Cleaned up some code. Some other misc stuff.

+++ src/dependencies/lcommon/Timer.cpp
@@ -53,7 +53,7 @@ void Timer::start() {
 	microseconds_since_epoch = timev.tv_sec * 1000 * 1000 + timev.tv_usec;
 }
 
-unsigned long Timer::get_microseconds() const {
+unsigned long long Timer::get_microseconds() const {
 	Timer now;
 	return now.microseconds_since_epoch - microseconds_since_epoch;
 }
